feat: Print logs to a file, change log level dynamically
