+++ .gitignore
@@ -12,3 +12,8 @@
 # env files
 /config/env/*
 !/config/env/.env.template
+
+#log files
+/logs
+/logs/all.log
+/logs/error.log
